fix: resolve FunctionOutputDTO circular dependency initialization error
- Move FunctionOutputDTO and related DTOs to separate file (Function/dto.ts)
- Break circular dependency chain between CommandService -> FunctionService -> EventService -> ModuleService
- Lazy-load queueService in CommandService to avoid initialization issues
- Re-export DTOs from FunctionService for backward compatibility

The circular dependency was causing "Cannot access 'FunctionOutputDTO' before initialization" error
because TypeScript decorators require immediate access to the type at module initialization time.
